+++ src/resolve.rs
@@ -16,7 +16,7 @@ pub trait Resolve: Sized {
     fn resolve<'a>(
         ctx: &'a RequestState,
         captures: &mut Captures,
-    ) -> ResolveGuard<Self::Output<'a>>;
+    ) -> Result<Self::Output<'a>, Box<dyn IntoResponseError>>;
 }
 
 /// `ResolveGuard` is the expected return type of top level `Resolve`able objects. Only types that
@@ -56,15 +56,15 @@ impl<K> Resolve for Query<'_, K>
 where
     K: TypeCacheKey,
 {
-    type Output<'b> = Query<'b, K>;
+    type Output<'a> = Query<'a, K>;
 
-    fn resolve<'c>(
-        ctx: &'c RequestState,
+    fn resolve<'b>(
+        ctx: &'b RequestState,
         _captures: &mut Captures,
-    ) -> ResolveGuard<Self::Output<'c>> {
+    ) -> Result<Self::Output<'b>, Box<dyn IntoResponseError>> {
         match ctx.global_cache.get::<K>().map(|v| Query(v)) {
-            Some(v) => ResolveGuard::Value(v),
-            None => ResolveGuard::err(Error::QueryNotInCache),
+            Some(v) => Ok(v),
+            None => Err(Box::new(Error::QueryNotInCache)),
         }
     }
 }
@@ -73,13 +73,13 @@ where
 pub struct Url<'a>(pub &'a str);
 
 impl Resolve for Url<'_> {
-    type Output<'b> = Url<'b>;
+    type Output<'a> = Url<'a>;
 
-    fn resolve<'c>(
-        ctx: &'c RequestState,
+    fn resolve<'b>(
+        ctx: &'b RequestState,
         _captures: &mut Captures,
-    ) -> ResolveGuard<Self::Output<'c>> {
-        ResolveGuard::Value(Url(ctx.request.uri().path()))
+    ) -> Result<Self::Output<'b>, Box<dyn IntoResponseError>> {
+        Ok(Url(ctx.request.uri().path()))
     }
 }
 
@@ -90,12 +90,15 @@ pub struct UrlPart(pub String);
 impl Resolve for UrlPart {
     type Output<'a> = Self;
 
-    fn resolve(_ctx: &RequestState, captures: &mut Captures) -> ResolveGuard<Self> {
+    fn resolve(
+        _ctx: &RequestState,
+        captures: &mut Captures,
+    ) -> Result<Self, Box<dyn IntoResponseError>> {
         let Some(part) = captures.pop_front() else {
-            return ResolveGuard::err(Error::MissingUrlPart);
+            return Err(Box::new(Error::MissingUrlPart));
         };
 
-        ResolveGuard::Value(UrlPart(part))
+        Ok(UrlPart(part))
     }
 }
 
@@ -106,12 +109,15 @@ pub struct UrlCollect(pub Vec<String>);
 impl Resolve for UrlCollect {
     type Output<'a> = Self;
 
-    fn resolve(_ctx: &RequestState, captures: &mut Captures) -> ResolveGuard<Self> {
+    fn resolve(
+        _ctx: &RequestState,
+        captures: &mut Captures,
+    ) -> Result<Self, Box<dyn IntoResponseError>> {
         let mut new = VecDeque::new();
 
         std::mem::swap(&mut new, captures);
 
-        ResolveGuard::Value(UrlCollect(Vec::from(new)))
+        Ok(UrlCollect(Vec::from(new)))
     }
 }
 
@@ -124,8 +130,8 @@ impl Resolve for HeaderMap<'_> {
     fn resolve<'c>(
         ctx: &'c RequestState,
         _captures: &mut Captures,
-    ) -> ResolveGuard<Self::Output<'c>> {
-        ResolveGuard::Value(HeaderMap(ctx.request.headers()))
+    ) -> Result<Self::Output<'c>, Box<dyn IntoResponseError>> {
+        Ok(HeaderMap(ctx.request.headers()))
     }
 }
 
@@ -138,8 +144,8 @@ impl Resolve for ArgMap<'_> {
     fn resolve<'c>(
         ctx: &'c RequestState,
         _captures: &mut Captures,
-    ) -> ResolveGuard<Self::Output<'c>> {
-        ResolveGuard::Value(ArgMap(&ctx.query))
+    ) -> Result<Self::Output<'c>, Box<dyn IntoResponseError>> {
+        Ok(ArgMap(&ctx.query))
     }
 }
 
@@ -152,10 +158,10 @@ where
     fn resolve<'c>(
         ctx: &'c RequestState,
         captures: &mut Captures,
-    ) -> ResolveGuard<Self::Output<'c>> {
+    ) -> Result<Self::Output<'c>, Box<dyn IntoResponseError>> {
         match T::resolve(ctx, captures) {
-            ResolveGuard::Value(v) => ResolveGuard::Value(Some(v)),
-            _ => ResolveGuard::Value(None),
+            Ok(v) => Ok(Some(v)),
+            _ => Ok(None),
         }
     }
 }
@@ -166,8 +172,8 @@ impl Resolve for &[u8] {
     fn resolve<'c>(
         ctx: &'c RequestState,
         _captures: &mut Captures,
-    ) -> ResolveGuard<Self::Output<'c>> {
-        ResolveGuard::Value(ctx.request.body().get_as_slice())
+    ) -> Result<Self::Output<'c>, Box<dyn IntoResponseError>> {
+        Ok(ctx.request.body().get_as_slice())
     }
 }
 
@@ -177,10 +183,10 @@ impl Resolve for &str {
     fn resolve<'c>(
         ctx: &'c RequestState,
         _captures: &mut Captures,
-    ) -> ResolveGuard<Self::Output<'c>> {
+    ) -> Result<Self::Output<'c>, Box<dyn IntoResponseError>> {
         match std::str::from_utf8(ctx.request.body().get_as_slice()) {
-            Ok(v) => ResolveGuard::Value(v),
-            Err(_) => ResolveGuard::err(Error::MalformedRequest),
+            Ok(v) => Ok(v),
+            Err(_) => Err(Box::new(Error::MalformedRequest)),
         }
     }
 }